refactor: 💡 relocate display components to shared folder
