+++ apps/webstore/src/app/shared/components/items/product-display/product-display.component.ts
@@ -1,5 +1,5 @@
 import { Component, Input } from '@angular/core';
-import { ProductsItem } from '../products-items.constant';
+import { ProductsItem } from '../../../../pages/products/products-items.constant';
 
 @Component({
   selector: 'webstore-product-display',
